refactor resource enpoint to a composable for getting the querystring addition
